pyproject.toml
[tool.poetry]
name = "vehicle-plate-validator-backend"
version = "0.1.0"
description = "Api backend for storing and retrieving plates numbers"
authors = ["rahulpoluri <rahulpoluri314@gmail.com>"]
readme = "README.md"
package-mode = false
packages = [
    { include = "application" }
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.5"}
uvicorn = "^0.32.1"
pydantic = "^2.10.2"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
httpx = "^0.28.0"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
mypy = "^1.13.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"
httpx = "^0.28.0"

[tool.poetry.scripts]
lint = "application.scripts.lint:main"

[tool.mypy]
plugins = ['pydantic.mypy']

[tool.black]
line-length = 79

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
